pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "lofi-gen"
version = "1.0.0"
description = "Professional Lofi Video Generator — CLI + API + Python SDK"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Antigravity AI"},
]
keywords = ["lofi", "video", "generator", "youtube", "ffmpeg", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "requests>=2.28.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "celery>=5.3.0",
    "redis>=4.6.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
]
ai = [
    "google-generativeai>=0.3.0",
]
media = [
    "moviepy>=1.0.3",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",
]
all = [
    "lofi-gen[api,ai,media]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
lofi-gen = "lofi_gen.cli:main"

[project.urls]
Homepage = "https://github.com/fhrrutedf/lofi-video-generator"
Repository = "https://github.com/fhrrutedf/lofi-video-generator"
Issues = "https://github.com/fhrrutedf/lofi-video-generator/issues"

[tool.setuptools.packages.find]
include = ["lofi_gen*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
